fix:コメント追加＆コードを修正
