+++ features/assets/components/portfolio-ai-summary.tsx
@@ -1,11 +1,10 @@
 "use client";
 
 // 智能资产分析助手
-// 它接收一个资产数组作为 props，调用后端 API 进行资产分析，并在组件中显示摘要内容
-import { useRef, useState } from "react";
-import type { Asset } from "@/types/global";
-import { useTranslations, useLocale } from "next-intl";
+// 逻辑层见 hooks/use-portfolio-ai-summary.ts
 
+import type { Asset } from "@/types/global";
+import { useTranslations } from "next-intl";
 import { Bot, Sparkles } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import {
@@ -16,62 +15,21 @@ import {
   DialogTrigger,
 } from "@/components/ui/dialog";
 import ReactMarkdown from "react-markdown";
+import { usePortfolioAISummary } from "@/features/assets/hooks/use-portfolio-ai-summary";
 
 type Props = { assets: Asset[] };
 
 export function PortfolioAISummary({ assets }: Props) {
   const t = useTranslations("aiSummary");
-  const locale = useLocale();
-  const [open, setOpen] = useState(false);
-  const [text, setText] = useState("");
-  const [loading, setLoading] = useState(false);
-  const abortRef = useRef<AbortController | null>(null);
 
-  // Dialogの開閉を制御 → 開いたときだけrun()を呼ぶ
-  function handleOpenChange(nextOpen: boolean) {
-    if (nextOpen) {
-      run(); // 開いた瞬間に分析開始
-    } else {
-      abortRef.current?.abort(); // 閉じたらキャンセル
-    }
-    setOpen(nextOpen);
-  }
-
-  async function run() {
-    if (assets.length === 0) return;
-    abortRef.current?.abort();
-    const ctrl = new AbortController();
-    abortRef.current = ctrl;
-
-    setText("");
-    setLoading(true);
-    try {
-      const res = await fetch("/api/assets/ai-summary", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ assets, locale }),
-        signal: ctrl.signal,
-      });
-      if (!res.ok || !res.body) throw new Error("fetch failed");
-
-      const reader = res.body.getReader();
-      const decoder = new TextDecoder();
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-        setText((prev) => prev + decoder.decode(value, { stream: true }));
-      }
-    } catch (e: unknown) {
-      if (e instanceof Error && e.name !== "AbortError") {
-        setText(t("fetchError"));
-      }
-    } finally {
-      setLoading(false);
-    }
-  }
+  const { open, handleOpenChange, text, loading } = usePortfolioAISummary({
+    assets,
+    fetchErrorMessage: t("fetchError"),
+  });
 
   return (
     <Dialog open={open} onOpenChange={handleOpenChange}>
+      {/* 触发区域：标题描述 + 分析按钮 */}
       <div className="px-3 sm:px-5 py-4 flex flex-col items-center justify-center gap-3 text-center">
         <div>
           <p className="text-[13px] font-medium">{t("title")}</p>
@@ -93,6 +51,7 @@ export function PortfolioAISummary({ assets }: Props) {
         </DialogTrigger>
       </div>
 
+      {/* 分析结果弹窗 */}
       <DialogContent className="w-[95vw] max-w-2xl">
         <DialogHeader>
           <DialogTitle className="flex items-center gap-2 text-base">
@@ -101,16 +60,12 @@ export function PortfolioAISummary({ assets }: Props) {
           </DialogTitle>
         </DialogHeader>
 
-        {/* ローディング or 結果表示 */}
+        {/* 加载中：骨架屏动画；加载完成：渲染 Markdown */}
         <div className="min-h-50 max-h-[60vh] overflow-y-auto pr-1">
           {loading ? (
             <div className="space-y-2.5 animate-pulse pt-2">
               {[75, 55, 85, 45, 65, 70, 50].map((w, i) => (
-                <div
-                  key={i}
-                  className="h-3 rounded bg-muted"
-                  style={{ width: `${w}%` }}
-                />
+                <div key={i} className="h-3 rounded bg-muted" style={{ width: `${w}%` }} />
               ))}
             </div>
           ) : text ? (
